+++ packages/hot-module-replacement/package.js
@@ -7,10 +7,13 @@ Package.describe({
 });
 
 Package.onUse(function (api) {
+  api.versionsFrom('2.2');
   api.use('modules');
   api.use('meteor');
+  api.use('hot-code-push', { unordered: true });
+
   api.use('dev-error-overlay', { weak: true });
-  api.imply('modules-runtime-hot@0.12.0');
+  api.imply('modules-runtime-hot@0.13.0');
   api.addFiles([
     './hot-api.js',
     './client.js'